requirements.txt
# Requirements for slack-annoyance
slack-sdk>=3.20.0
requests>=2.28.0
python-dotenv>=0.21.0
langfuse>=0.1.0
pinecone>=2.2.1
flask>=2.0.0
